requirements.txt
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
google-genai
feedparser
gTTS
moviepy
Pillow
python-dotenv
requests
anthropic
